Put all library files into a separate folder
