+++ src/python/janelia_emrp/msem/wafer_60_gc_upload/details/config.py
